consolidation of examples: date_time_functions  (#14240)
* consolidation: date_time_functions

* Documentation
