+++ datafusion-examples/examples/dataframe.rs
@@ -64,7 +64,6 @@ async fn main() -> Result<()> {
     read_csv(&ctx).await?;
     read_memory(&ctx).await?;
     write_out(&ctx).await?;
-    query_to_date().await?;
     register_aggregate_test_data("t1", &ctx).await?;
     register_aggregate_test_data("t2", &ctx).await?;
     where_scalar_subquery(&ctx).await?;
@@ -227,41 +226,6 @@ async fn write_out(ctx: &SessionContext) -> std::result::Result<(), DataFusionEr
             None,
         )
         .await?;
-
-    Ok(())
-}
-
-/// This example demonstrates how to use the to_date series
-/// of functions in the DataFrame API as well as via sql.
-async fn query_to_date() -> Result<()> {
-    // define a schema.
-    let schema = Arc::new(Schema::new(vec![Field::new("a", DataType::Utf8, false)]));
-
-    // define data.
-    let batch = RecordBatch::try_new(
-        schema,
-        vec![Arc::new(StringArray::from(vec![
-            "2020-09-08T13:42:29Z",
-            "2020-09-08T13:42:29.190855-05:00",
-            "2020-08-09 12:13:29",
-            "2020-01-02",
-        ]))],
-    )?;
-
-    // declare a new context. In spark API, this corresponds to a new spark SQLsession
-    let ctx = SessionContext::new();
-
-    // declare a table in memory. In spark API, this corresponds to createDataFrame(...).
-    ctx.register_batch("t", batch)?;
-    let df = ctx.table("t").await?;
-
-    // use to_date function to convert col 'a' to timestamp type using the default parsing
-    let df = df.with_column("a", to_date(vec![col("a")]))?;
-
-    let df = df.select_columns(&["a"])?;
-
-    // print the results
-    df.show().await?;
 
     Ok(())
 }
